Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main(){
-    int x,y,z;
-    scanf("%d",&x);
-    scanf("%d",&y);
-    scanf("%d",&z);
+    float x,y,z;
+    scanf("%f",&x);
+    scanf("%f",&y);
+    scanf("%f",&z);
     float average = (x+y+z)/3;
     printf("Average: %.2f",average);
 }
